Complete update, move, and reformat of lib engine and install generator
