+++ lib/generators/spree_address_book/install/install_generator.rb
@@ -1,0 +1,38 @@
+module SpreeAddressBook
+  module Generators
+    class InstallGenerator < Rails::Generators::Base
+
+      class_option :auto_run_migrations, :type => :boolean, :default => false
+      source_root File.expand_path("../../../templates",__FILE__)
+
+      def add_javascripts
+        append_file 'vendor/assets/javascripts/spree/frontend/all.js', "//= require spree/frontend/\n"
+        append_file 'vendor/assets/javascripts/spree/frontend/all.js', "//= require spree/frontend/spree_address_book\n"
+        append_file 'vendor/assets/javascripts/spree/backend/all.js', "//= require spree/backend/spree_address_book\n"
+      end
+
+      def add_stylesheets
+        inject_into_file 'vendor/assets/stylesheets/spree/frontend/all.css', " *= require spree/frontend/spree_address_book\n", :before => /\*\//, :verbose => true
+        inject_into_file 'vendor/assets/stylesheets/spree/backend/all.css', " *= require spree/backend/spree_address_book\n", :before => /\*\//, :verbose => true
+      end
+
+      def add_migrations
+        run 'bundle exec rake railties:install:migrations FROM=spree_address_book'
+      end
+
+      def add_initializer_file
+        copy_file "spree_address_book.rb", "config/initializers/spree_address_book.rb"
+      end
+
+      def run_migrations
+        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask 'Would you like to run the migrations now? [Y/n]')
+        res = ask "Would you like to run the migrations now? [Y/n]"
+        if run_migrations
+          run 'bundle exec rake db:migrate'
+        else
+          puts 'Skipping rake db:migrate, don\'t forget to run it!'
+        end
+      end
+    end
+  end
+end
