--- lib/generators/spree_address_book/install_generator.rb
@@ -1,26 +1,0 @@
-module SpreeAddressBook
-  module Generators
-    class InstallGenerator < Rails::Generators::Base
-      def add_javascripts
-        append_file "app/assets/javascripts/store/all.js", "//= require store/spree_address_book\n"
-      end
-      
-      def add_stylesheets
-        inject_into_file "app/assets/stylesheets/store/all.css", " *= require store/spree_address_book\n", :before => /\*\//, :verbose => true
-      end
-    
-      def add_migrations
-        run 'bundle exec rake railties:install:migrations FROM=spree_address_book'
-      end
-
-      def run_migrations
-         res = ask "Would you like to run the migrations now? [Y/n]"
-         if res == "" || res.downcase == "y"
-           run 'bundle exec rake db:migrate'
-         else
-           puts "Skiping rake db:migrate, don't forget to run it!"
-         end
-      end
-    end
-  end
-end
