+++ lib/generators/templates/spree_address_book.rb
@@ -1,0 +1,3 @@
+Spree.config do |config|
+  config.address_requires_state = true
+end
