+++ lib/spree_address_book/factories.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  # Define your Spree extensions Factories within this file to enable applications, and other extensions to use and override them.
+  #
+  # Example adding this to your spec_helper will load these Factories for use:
+  # require 'spree_address_book/factories'
+end
